+++ client/shared/src/settings/temporary/TemporarySettings.ts
@@ -15,6 +15,8 @@ export interface TemporarySettingsSchema {
     'search.sidebar.revisions.tab': number
     'search.onboarding.tourCancelled': boolean
     'search.contexts.ctaDismissed': boolean
+    'search.notepad.enabled': boolean
+    'search.notepad.ctaSeen': boolean
     'insights.freeGaAccepted': boolean
     'insights.freeGaExpiredAccepted': boolean
     'insights.wasMainPageOpen': boolean
@@ -34,7 +36,7 @@ export interface TemporarySettingsSchema {
 }
 
 /**
- * All temporary setttings are possibly undefined. This is the actual schema that
+ * All temporary settings are possibly undefined. This is the actual schema that
  * should be used to force the consumer to check for undefined values.
  */
 export type TemporarySettings = Optional<TemporarySettingsSchema>