+++ client/web/src/enterprise/insights/core/backend/gql-backend/code-insights-gql-backend.ts
@@ -112,10 +112,11 @@ export class CodeInsightsGqlBackend implements CodeInsightsBackend {
         fromObservableQuery(
             this.apolloClient.watchQuery<HasAvailableCodeInsightResult>({
                 query: gql`
-                    query HasAvailableCodeInsight($count: Int!) {
-                        insightViews(first: $count) {
+                    query HasAvailableCodeInsight {
+                        insightViews {
                             nodes {
                                 id
+                                isFrozen
                             }
                         }
                     }
@@ -123,7 +124,7 @@ export class CodeInsightsGqlBackend implements CodeInsightsBackend {
                 variables: { count: insightsCount },
                 nextFetchPolicy: 'cache-only',
             })
-        ).pipe(map(({ data }) => data.insightViews.nodes.length === insightsCount))
+        ).pipe(map(({ data }) => data.insightViews.nodes.filter(node => !node.isFrozen).length === insightsCount))
 
     // TODO: This method is used only for insight title validation but since we don't have
     // limitations about title field in gql api remove this method and async validation for
@@ -229,53 +230,14 @@ export class CodeInsightsGqlBackend implements CodeInsightsBackend {
 
     // Live preview fetchers
     public getSearchInsightContent = (input: GetSearchInsightContentInput): Promise<SeriesChartContent<any>> =>
-        getSearchInsightContent(input).then(data => {
-            const { data: datumList, series, xAxis } = data
-
-            // TODO: Remove this when the dashboard page has new chart fetchers
-            return {
-                data: datumList,
-                series: series.map(series => ({
-                    dataKey: series.dataKey,
-                    name: series.name ?? '',
-                    color: series.stroke,
-                    getLinkURL: datum => series.linkURLs?.[+datum[xAxis.dataKey]] ?? undefined,
-                })),
-                getXValue: datum => new Date(+datum[xAxis.dataKey]),
-            }
-        })
+        getSearchInsightContent(input).then(data => data.content)
 
     public getLangStatsInsightContent = (
         input: GetLangStatsInsightContentInput
-    ): Promise<CategoricalChartContent<any>> =>
-        getLangStatsInsightContent(input).then(data => {
-            const { data: dataList, dataKey, nameKey, fillKey = '', linkURLKey = '' } = data.pies[0]
-
-            // TODO: Remove this when the dashboard page has new chart fetchers
-            return {
-                data: dataList,
-                getDatumValue: datum => datum[dataKey],
-                getDatumColor: datum => datum[fillKey ?? ''],
-                getDatumName: datum => datum[nameKey],
-                getDatumLink: datum => datum[linkURLKey],
-            }
-        })
+    ): Promise<CategoricalChartContent<any>> => getLangStatsInsightContent(input).then(data => data.content)
 
     public getCaptureInsightContent = (input: CaptureInsightSettings): Promise<SeriesChartContent<any>> =>
-        getCaptureGroupInsightsPreview(this.apolloClient, input).then(data => {
-            const { data: datumList, series, xAxis } = data
-
-            // TODO: Remove this when the dashboard page has new chart fetchers
-            return {
-                data: datumList,
-                series: series.map(series => ({
-                    dataKey: series.dataKey,
-                    name: series.name ?? '',
-                    color: series.stroke,
-                })),
-                getXValue: datum => new Date(+datum[xAxis.dataKey]),
-            }
-        })
+        getCaptureGroupInsightsPreview(this.apolloClient, input)
 
     // Repositories API
     public getRepositorySuggestions = getRepositorySuggestions