+++ client/web/src/notebooks/notebookPage/NotebookPage.tsx
@@ -1,8 +1,8 @@
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 
 import classNames from 'classnames'
+import BookOutlineIcon from 'mdi-react/BookOutlineIcon'
 import CheckCircleIcon from 'mdi-react/CheckCircleIcon'
-import MagnifyIcon from 'mdi-react/MagnifyIcon'
 import { RouteComponentProps } from 'react-router'
 import { Observable } from 'rxjs'
 import { catchError, delay, startWith, switchMap } from 'rxjs/operators'
@@ -77,7 +77,6 @@ export const NotebookPage: React.FunctionComponent<NotebookPageProps> = ({
     telemetryService,
     searchContextsEnabled,
     isSourcegraphDotCom,
-    location,
     fetchHighlightedFileLineRanges,
     authenticatedUser,
     showSearchContext,
@@ -199,8 +198,7 @@ export const NotebookPage: React.FunctionComponent<NotebookPageProps> = ({
                             <PageHeader
                                 className="mt-2"
                                 path={[
-                                    { icon: MagnifyIcon, to: '/search' },
-                                    { to: '/notebooks', text: 'Notebooks' },
+                                    { to: '/notebooks', icon: BookOutlineIcon, ariaLabel: 'Notebooks' },
                                     {
                                         text: (
                                             <NotebookTitle
@@ -283,7 +281,6 @@ export const NotebookPage: React.FunctionComponent<NotebookPageProps> = ({
                                 telemetryService={telemetryService}
                                 searchContextsEnabled={searchContextsEnabled}
                                 isSourcegraphDotCom={isSourcegraphDotCom}
-                                location={location}
                                 fetchHighlightedFileLineRanges={fetchHighlightedFileLineRanges}
                                 authenticatedUser={authenticatedUser}
                                 showSearchContext={showSearchContext}