+++ internal/database/migration/cliutil/downto.go
@@ -1,43 +1,55 @@
 package cliutil
 
 import (
-	"context"
 	"flag"
 	"fmt"
 	"strconv"
 	"strings"
 
-	"github.com/peterbourgon/ff/v3/ffcli"
+	"github.com/urfave/cli/v2"
 
 	"github.com/sourcegraph/sourcegraph/internal/database/migration/runner"
 	"github.com/sourcegraph/sourcegraph/lib/output"
 )
 
-func DownTo(commandName string, factory RunnerFactory, out *output.Output, development bool) *ffcli.Command {
-	var (
-		flagSet                  = flag.NewFlagSet(fmt.Sprintf("%s downto", commandName), flag.ExitOnError)
-		schemaNameFlag           = flagSet.String("db", "", `The target schema to modify.`)
-		unprivilegedOnlyFlag     = flagSet.Bool("unprivileged-only", false, `Do not apply privileged migrations.`)
-		ignoreSingleDirtyLogFlag = flagSet.Bool("ignore-single-dirty-log", development, `Ignore a previously failed attempt if it will be immediately retried by this operation.`)
-		targetsFlag              = flagSet.String("target", "", "Revert all children of the given target. Comma-separated values are accepted.")
-	)
+func DownTo(commandName string, factory RunnerFactory, out *output.Output, development bool) *cli.Command {
+	flags := []cli.Flag{
+		&cli.StringFlag{
+			Name:     "db",
+			Usage:    "The target `schema` to modify.",
+			Required: true,
+		},
+		&cli.StringFlag{
+			Name:     "target",
+			Usage:    `The migration to apply. Comma-separated values are accepted.`,
+			Required: true,
+		},
+		&cli.BoolFlag{
+			Name:  "unprivileged-only",
+			Usage: `Do not apply privileged migrations.`,
+			Value: false,
+		},
+		&cli.BoolFlag{
+			Name:  "ignore-single-dirty-log",
+			Usage: `Ignore a previously failed attempt if it will be immediately retried by this operation.`,
+			Value: development,
+		},
+	}
 
-	exec := func(ctx context.Context, args []string) error {
-		if len(args) != 0 {
+	action := func(cmd *cli.Context) error {
+		if cmd.NArg() != 0 {
 			out.WriteLine(output.Linef("", output.StyleWarning, "ERROR: too many arguments"))
 			return flag.ErrHelp
 		}
 
-		if *schemaNameFlag == "" {
-			out.WriteLine(output.Linef("", output.StyleWarning, "ERROR: supply a schema via -db"))
-			return flag.ErrHelp
-		}
+		var (
+			schemaNameFlag           = cmd.String("db")
+			unprivilegedOnlyFlag     = cmd.Bool("unprivileged-only")
+			ignoreSingleDirtyLogFlag = cmd.Bool("ignore-single-dirty-log")
+			targetsFlag              = cmd.String("target")
+		)
 
-		if *targetsFlag == "" {
-			out.WriteLine(output.Linef("", output.StyleWarning, "ERROR: supply a migration target via -target"))
-			return flag.ErrHelp
-		}
-		targets := strings.Split(*targetsFlag, ",")
+		targets := strings.Split(targetsFlag, ",")
 
 		versions := make([]int, 0, len(targets))
 		for _, target := range targets {
@@ -49,7 +61,8 @@ func DownTo(commandName string, factory RunnerFactory, out *output.Output, devel
 			versions = append(versions, version)
 		}
 
-		r, err := factory(ctx, []string{*schemaNameFlag})
+		ctx := cmd.Context
+		r, err := factory(ctx, []string{schemaNameFlag})
 		if err != nil {
 			return err
 		}
@@ -57,22 +70,22 @@ func DownTo(commandName string, factory RunnerFactory, out *output.Output, devel
 		return r.Run(ctx, runner.Options{
 			Operations: []runner.MigrationOperation{
 				{
-					SchemaName:     *schemaNameFlag,
+					SchemaName:     schemaNameFlag,
 					Type:           runner.MigrationOperationTypeTargetedDown,
 					TargetVersions: versions,
 				},
 			},
-			UnprivilegedOnly:     *unprivilegedOnlyFlag,
-			IgnoreSingleDirtyLog: *ignoreSingleDirtyLogFlag,
+			UnprivilegedOnly:     unprivilegedOnlyFlag,
+			IgnoreSingleDirtyLog: ignoreSingleDirtyLogFlag,
 		})
 	}
 
-	return &ffcli.Command{
-		Name:       "downto",
-		ShortUsage: fmt.Sprintf("%s downto -db=<schema> -target=<target>,<target>,...", commandName),
-		ShortHelp:  `Revert any applied migrations that are children of the given targets - this effectively "resets" the schmea to the target version`,
-		FlagSet:    flagSet,
-		Exec:       exec,
-		LongHelp:   ConstructLongHelp(),
+	return &cli.Command{
+		Name:        "downto",
+		UsageText:   fmt.Sprintf("%s downto -db=<schema> -target=<target>,<target>,...", commandName),
+		Usage:       `Revert any applied migrations that are children of the given targets - this effectively "resets" the schmea to the target version`,
+		Description: ConstructLongHelp(),
+		Flags:       flags,
+		Action:      action,
 	}
 }